# Prevent degenerate case in `DynamicBVH`