@lawnjelly 2025-08-12
Just for a historical note, the adding everything to the origin in the BVH, then moving paradigm is something we noticed as problematic a while ago (circa 2020?) around when I was writing the BVH used for physics and rendering in 3.x (this was since replaced for rendering in 4.x by Juan's "dynamic BVH"). PouleyKetchoup did a fair bit of research into this if I remember right when he was maintaining physics but some of these issues / PRs may have been deleted (can't find them now by searching, but will post if I can find).

The issue was not just adding to the BVH itself, it was that you would get pairings of every combination at the origin (which scales with `O(N squared)` or something like that). Pairings would occur and affect lighting, shadows, collision detection, etc etc.

As far as processing the better solution would be add the object, move to initial location, _then_ activate in the BVH. But this had compat breaking implications. Pouley left the project before we fixed up any of this.