@precup 2025-08-14
> Is this code available somewhere so I can run it locally? (You could upload it to a separate branch, for instance.)

I pushed it to [precup:bvh-benchmark](https://github.com/precup/godot/tree/bvh-benchmark). It's ugly, sorry. It's also set up a bit differently than the one I originally used because I had been using it to test other options, but the core benchmark is the same. It's currently set up to get numbers on `insert` time changes. If you want to see evidence of the bad case, uncomment the other benchmark calls and drop the trials to like 10.

I'll note for posterity in case anyone's curious: I had more recently tried two other solutions to get the cost lower, and neither worked. One was adding `unlikely` to the condition, which slowed inserts down by an additional 1%, and the other was using `fpclassify` to try to cheapen both comparisons, which added an additional 2%.